--- src/Views/Home/Illustrations/NoEventIllustration.tsx
@@ -1,217 +1,0 @@
-import Svg, { Path } from "react-native-svg";
-
-export interface NoEventIllustrationProps {
-  width?: string;
-  height?: string;
-}
-
-export default function NoEventIllustration({
-  width = "100%",
-  height = "100%"
-}: NoEventIllustrationProps): JSX.Element {
-  return (
-    <Svg width={width} height={height} viewBox="0 0 415 254">
-      <Path
-        opacity=".3"
-        d="M89.07 20.73C49.798 30.11 18.774 51.96 13.8 92.107c-4.187 33.819 5.483 91.304 72.122 97.422 134.712 12.364 278.227 24.752 297.637-33.542 13.163-39.522 17.501-110.238-55.199-131.97C279.05 9.255 173.244.64 89.069 20.73Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        opacity=".45"
-        d="M205.829 241.599c100.681 0 182.3-6.99 182.3-15.614s-81.619-15.615-182.3-15.615c-100.681 0-182.3 6.991-182.3 15.615s81.619 15.614 182.3 15.614Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M179.942 141.355a4.256 4.256 0 0 0 3.076-.266 5.788 5.788 0 0 0 2.43-1.804c.633-.819.92-1.853.798-2.88a1.526 1.526 0 0 0-.648-1.157c-.301-.174-.694-.185-.983-.37a3.022 3.022 0 0 1-.706-.902 3.022 3.022 0 0 0-2.313-1.157 5.961 5.961 0 0 0-2.626.578 4.252 4.252 0 0 0-2.614 2.499 3.464 3.464 0 0 0 1.503 3.469 10.405 10.405 0 0 0 3.783 1.435"
-        fill="#24285B"
-      />
-      <Path
-        d="M156.911 132.969a3.478 3.478 0 0 1 3.17-1.052c.375.111.766.158 1.156.139.308-.085.604-.209.879-.37 1.504-.694 3.262-.07 4.801.566a3.971 3.971 0 0 0 1.631.417c.478-.04.947-.158 1.388-.347a18.342 18.342 0 0 1 3.239-.81 2.038 2.038 0 0 1 1.573.22c.215.202.416.418.601.648.442.362.982.583 1.55.636a5.858 5.858 0 0 1 3.262 1.156c.301.301.521.671.81.96.289.289.578.474.833.752a4.156 4.156 0 0 1 .59 3.146 1.888 1.888 0 0 1-.243.902 1.829 1.829 0 0 1-1.273.579l-6.338 1.017c-4.211.671-8.421 1.354-12.666 1.863-2.146.34-4.326.41-6.49.208-2.313-.301-2.938-1.284-3.134-3.47-.313-3.366 2.718-5.055 4.661-7.16Z"
-        fill="#E6E6E6"
-      />
-      <Path d="M188.247 136.439h-37.328v15.95h37.328v-15.95Z" fill="#FFD200" />
-      <Path
-        d="m119.792 134.519 3.031-7.668 6.697-16.968 6.802-17.268c1.758-4.43 3.539-8.663 5.702-12.92a16.408 16.408 0 0 1 2.499-3.944 5.51 5.51 0 0 1 4.118-1.92c.2-.013.399.04.566.15a.977.977 0 0 1 .278.428 8.983 8.983 0 0 1 .15 5.703 31.433 31.433 0 0 1-1.526 4.117 10.871 10.871 0 0 0-.729 3.285c0 .139-.058.393-.081.555l-.59 3.944-2.314 15.279-2.625 17.627a17496.3 17496.3 0 0 1-1.643 11.011c0 .162 1.157-5.91-.092.683-1.25 6.592-12.805 18.228-20.243-2.094Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M184.499 177.314s10.145 9.554-2.996 9.172c-13.14-.381-8.305-10.606 2.996-9.172Z"
-        fill="#55789B"
-      />
-      <Path
-        d="M108.768 69.84s19.665 14.702 10.411 20.369-16.414-21.733-16.414-21.733l6.003 1.365Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M108.676 57.315s1.48 6.222 1.029 10.583a3.19 3.19 0 0 1-1.179 2.17 3.202 3.202 0 0 1-2.372.686 7.116 7.116 0 0 1-6.27-4.476l-2.706-5.262a5.783 5.783 0 0 1 1.62-6.443c3.192-3.088 9.172-1.11 9.878 2.742Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="m97.71 64.278 1.967 20.888 11.44-1.908-6.027-14.435-7.38-4.545Zm5.032-.556s-.405-2.429-2.21-1.977c-1.804.45-1.237 3.978 1.296 3.944l.914-1.967Zm6.639-2.44 2.129 2.186a1.036 1.036 0 0 1 .261.955 1.04 1.04 0 0 1-.666.733l-2.313.81.589-4.684Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M146.674 160.219s39.132-9.923 44.233 7.368c5.101 17.292-54.713 15.591-54.713 15.591L122.58 172.41l24.094-12.191Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M106.328 82.217s22.776 2.313 38.646 20.46c9.427 10.769 5.864 19.744 1.411 25.261a18.126 18.126 0 0 0-2.973 16.991l4.72 14.7s-25.194 16.193-41.11 14.528c-15.917-1.666-66.003-89.442-.694-91.94Z"
-        fill="#55789B"
-      />
-      <Path
-        opacity=".08"
-        d="M75.304 100.653s-2.822 8.64-1.284 16.991c1.539 8.351 9.416 63.811 61.642 48.833l12.47-6.94-13.233-11.138-25.321-26.487-11.081-21.27-23.193.011Z"
-      />
-      <Path
-        d="M93.187 138.128c-11.197-33.912-61.179-28.361-65.459 7.102-4.407 36.607.59 72.416 59.525 78.974 93.695 10.409 108.478-17.222 95.256-39.522-13.221-22.3-79.004-10.583-88.824-44.97a18.312 18.312 0 0 0-.498-1.584Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M170.491 157.235s13.071 3.169 14.008-5.147l-14.008-1.515v6.662Zm-79.432-55.633c-.139-7.854 8.652-12.908 15.199-8.582 4.488 2.96 8.907 9.056 11.568 21.039 5.205 23.41 18.16 33.542 52.665 36.514v6.662s-78.032 19.836-79.432-55.633Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M107.38 58.83a4.627 4.627 0 0 0 1.157-7.923 9.717 9.717 0 0 0-2.313-1.33c-5.449-2.313-18.508-.683-17.64 11.705.867 12.387-19.815 40.551-1.157 48.196 18.658 7.646 34.216-2.452 34.216-11.566s-17.721-11.254-18.878-19.188a636.778 636.778 0 0 1-1.307-9.589 9.895 9.895 0 0 1 5.922-10.305Z"
-        fill="#24285B"
-      />
-      <Path
-        opacity=".45"
-        d="M61.956 176.667c-6.594-6.015-16.923.578-14.078 9.033 4.037 11.959 13.314 23.687 33.696 27.157 45.794 7.796 41.33-9.149 38.368-16.539-2.463-6.038-32.746 3.377-57.986-19.651Z"
-        fill="#fff"
-      />
-      <Path
-        d="M150.387 63.445s-1.399-5.598 3.204-7.773l.602 1.643s-2.522.092-2.915 6.268l-.891-.138Z"
-        fill="#55789B"
-      />
-      <Path
-        opacity=".08"
-        d="M150.387 63.445s-1.399-5.598 3.204-7.773l.602 1.643s-2.522.092-2.915 6.268l-.891-.138Z"
-      />
-      <Path
-        d="M146.663 78.758a5.114 5.114 0 0 1 3.285.405l.358.174a3.16 3.16 0 0 0 3.204-.243 14.303 14.303 0 0 0 6.362-10.167c.556-5.991-3.042-7.206-6.211-6.454a5.958 5.958 0 0 1-4.5-.728 5.308 5.308 0 0 0-4.916-.706c-3.204 1.411-4.916 11.416-.162 16.864a2.657 2.657 0 0 0 2.58.855Z"
-        fill="#FFD200"
-      />
-      <Path
-        opacity=".43"
-        d="M157.964 66.649c.083 2.5-.551 4.972-1.828 7.124a.748.748 0 0 1-.841.323.756.756 0 0 1-.547-.716v-6.674a1.598 1.598 0 0 1 2.117-1.526 1.572 1.572 0 0 1 1.099 1.469Z"
-        fill="#fff"
-      />
-      <Path
-        d="M150.85 58.332s.15-5.91-6.281-4.626c0-.058.069 5.933 6.281 4.626Z"
-        fill="#55789B"
-      />
-      <Path
-        opacity=".46"
-        d="M150.85 58.332s.15-5.91-6.281-4.626c0-.058.069 5.933 6.281 4.626Z"
-        fill="#fff"
-      />
-      <Path
-        d="M273.37 142.431c-5.356 6.338-12.25 11.658-20.289 13.706a4.627 4.627 0 0 0-2.915 1.492 3.932 3.932 0 0 0 .093 3.55 23.938 23.938 0 0 0 25.448 14.574 32.238 32.238 0 0 0 12.018-4.974 26.809 26.809 0 0 0 8.629-8.189c2.591-4.221 3.32-9.31 3.598-14.249a95.787 95.787 0 0 0-.174-13.382c-.416-4.904-3.979-16.922-11.475-14.539-3.088.972-5.135 7.518-6.674 10.12a84.789 84.789 0 0 1-8.259 11.891Z"
-        fill="#55789B"
-      />
-      <Path
-        opacity=".08"
-        d="M273.37 142.431c-5.356 6.338-12.25 11.658-20.289 13.706a4.627 4.627 0 0 0-2.915 1.492 3.932 3.932 0 0 0 .093 3.55 23.938 23.938 0 0 0 25.448 14.574 32.238 32.238 0 0 0 12.018-4.974 26.809 26.809 0 0 0 8.629-8.189c2.591-4.221 3.32-9.31 3.598-14.249a95.787 95.787 0 0 0-.174-13.382c-.416-4.904-3.979-16.922-11.475-14.539-3.088.972-5.135 7.518-6.674 10.12a84.789 84.789 0 0 1-8.259 11.891Z"
-      />
-      <Path
-        d="M264.463 150.041a4.28 4.28 0 0 0 3.089-.266 5.785 5.785 0 0 0 2.429-1.804 3.862 3.862 0 0 0 .786-2.88 1.537 1.537 0 0 0-.647-1.157c-.301-.173-.683-.185-.983-.37a2.965 2.965 0 0 1-.694-.902 3.007 3.007 0 0 0-2.314-1.157 5.867 5.867 0 0 0-2.614.579 4.215 4.215 0 0 0-2.603 2.498 3.476 3.476 0 0 0 1.516 3.47 10.222 10.222 0 0 0 3.77 1.434"
-        fill="#24285B"
-      />
-      <Path
-        d="M241.468 141.656a3.523 3.523 0 0 1 3.169-1.053c.375.11.766.157 1.157.139.302-.086.593-.206.867-.359 1.516-.694 3.274-.069 4.812.555a4.069 4.069 0 0 0 1.631.428c.475-.045.94-.166 1.377-.358a17.42 17.42 0 0 1 3.239-.798 2.035 2.035 0 0 1 1.573.208c.218.199.419.415.601.648a2.86 2.86 0 0 0 1.562.636 5.884 5.884 0 0 1 3.25 1.156c.301.301.521.671.81.972.289.301.59.463.833.74.302.461.51.978.612 1.52a4.181 4.181 0 0 1-.022 1.638c.001.312-.078.619-.232.89a1.85 1.85 0 0 1-1.272.579l-6.339 1.018c-4.222.682-8.432 1.353-12.666 1.873-2.15.337-4.334.403-6.501.197-2.313-.301-2.938-1.284-3.123-3.47-.301-3.354 2.718-5.043 4.662-7.159Z"
-        fill="#E6E6E6"
-      />
-      <Path d="M272.803 145.137h-37.327v15.95h37.327v-15.95Z" fill="#FFD200" />
-      <Path
-        d="M233.984 100.468h-11.012a1.383 1.383 0 0 0-1.377 1.446l.902 18.182a2.468 2.468 0 0 0 2.464 2.313h7.207a2.463 2.463 0 0 0 2.463-2.371l.729-18.17a1.386 1.386 0 0 0-1.376-1.4Z"
-        fill="#55789B"
-      />
-      <Path
-        d="m230.837 117.563-.809-.058 1.434-20.75a.409.409 0 0 1 .15-.289.464.464 0 0 1 .324-.08l6.675 1.156-.128.798-6.246-1.052-1.4 20.275Z"
-        fill="#24285B"
-      />
-      <Path
-        opacity=".44"
-        d="M233.602 107.107h-10.156l.451 12.376a1.541 1.541 0 0 0 1.538 1.481h6.096a1.541 1.541 0 0 0 1.539-1.469l.532-12.388Z"
-        fill="#fff"
-      />
-      <Path
-        d="M298.448 55.869s-.174 7.31 1.353 12.075a3.681 3.681 0 0 0 4.627 2.313 8.096 8.096 0 0 0 5.946-6.5l1.781-6.512a6.567 6.567 0 0 0-3.331-6.824c-4.28-2.648-10.492.96-10.376 5.448Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="m311.854 58.448 5.263 23.399-13.105 1.977 1.434-17.823 6.408-7.553Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M300.241 56.03a29.482 29.482 0 0 0 6.697-.936 6.07 6.07 0 0 0 .428 6.673 4.952 4.952 0 0 0 6.108 1.608l-.706-9.38a7.471 7.471 0 0 0-4.187-6.396 28.775 28.775 0 0 0-3.239-1.33c-2.765-.914-6.605 1.156-9.15-.532a1.77 1.77 0 0 0-2.718 1.33c-.243 2.88.555 7.344 4.511 8.663a7.085 7.085 0 0 0 2.256.3Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M306.603 61.652s-.127-2.822 2.001-2.73c2.128.093 2.313 4.141-.521 4.708l-1.48-1.978Zm-8.028-1.157-1.804 2.904a1.154 1.154 0 0 0 .301 1.59c.164.114.355.184.555.202l2.822.347-1.874-5.043Z"
-        fill="#F4A28C"
-      />
-      <Path
-        opacity=".31"
-        d="M304.972 69.887a8.653 8.653 0 0 0 4.106-2.81s.093 4.383-4.442 9.01l.336-6.2Z"
-        fill="#CE8172"
-      />
-      <Path
-        d="m323.479 174.342-35.418-2.314s-.509-3.469-1.284-9.114c-.556-4.059-1.261-9.253-2.013-14.92-1.238-9.403-2.603-20.31-3.69-30.327-2.672-24.486 22.938-33.866 22.938-33.866l13.105-1.978c59.549 9.254 6.362 92.519 6.362 92.519Z"
-        fill="#55789B"
-      />
-      <Path
-        opacity=".08"
-        d="M320.46 138.591c-16.506 20.09-21.526 32.281-21.526 32.281l-12.157-7.935c-.556-4.059-1.261-9.253-2.013-14.92l9.693-14.307 28.34-31.808.891.209s13.279 16.389-3.228 36.48Z"
-      />
-      <Path
-        d="M322.519 96.455a12.084 12.084 0 0 0-13.88-8.606c-6.732 1.296-14.76 5.783-17.583 19.466a103.77 103.77 0 0 1-11.856 31.437 7.492 7.492 0 0 1-11.382 1.863l-28.178-24.29-6.316 7.414 32.944 34.41a13.614 13.614 0 0 0 17.119 2.093c16.53-10.467 47.519-34.49 39.132-63.787Z"
-        fill="#55789B"
-      />
-      <Path
-        opacity=".46"
-        d="M322.519 96.455a12.084 12.084 0 0 0-13.88-8.606c-6.732 1.296-14.76 5.783-17.583 19.466a103.77 103.77 0 0 1-11.856 31.437 7.492 7.492 0 0 1-11.382 1.863l-28.178-24.29-6.316 7.414 32.944 34.41a13.614 13.614 0 0 0 17.119 2.093c16.53-10.467 47.519-34.49 39.132-63.787Z"
-        fill="#fff"
-      />
-      <Path
-        d="M239.629 116.302s-8.641-9.114-13.129-2.857c-4.488 6.258 6.813 10.271 6.813 10.271l6.316-7.414Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M242.335 177.314s-10.144 9.554 3.008 9.172c13.152-.381 8.305-10.606-3.008-9.172Z"
-        fill="#55789B"
-      />
-      <Path
-        d="M284.544 162.44s-43.504-12.145-48.582 5.147c-5.078 17.292 54.713 15.591 54.713 15.591l13.603-10.768-19.734-9.97Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M318.448 138.128c11.197-33.912 61.19-28.361 65.459 7.102 4.407 36.607-.59 72.416-59.525 78.974-93.695 10.409-108.478-17.222-95.257-39.522 13.222-22.3 79.005-10.583 88.837-44.97.15-.566.312-1.075.486-1.584Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        opacity=".45"
-        d="M349.702 176.667c6.594-6.015 16.923.578 14.078 9.033-4.026 11.959-13.314 23.687-33.696 27.157-45.783 7.796-41.318-9.149-38.368-16.539 2.406-6.038 32.735 3.377 57.986-19.651Z"
-        fill="#fff"
-      />
-      <Path
-        opacity=".58"
-        d="M59.075 211.62s-11.335-3.1-13.8-13.648c0 0 17.56-3.54 18.057 14.585l-4.257-.937Z"
-        fill="#55789B"
-      />
-      <Path
-        opacity=".73"
-        d="M60.464 210.498s-7.924-12.526-.949-24.22c0 0 13.349 8.478 7.415 24.289l-6.466-.069Z"
-        fill="#55789B"
-      />
-      <Path
-        d="M62.5 210.509s4.187-13.22 16.807-15.73c0 0 2.313 8.582-8.19 15.765l-8.618-.035Z"
-        fill="#55789B"
-      />
-      <Path
-        d="m54.287 210.209 2.29 15.718 14.47.058 2.14-15.695-18.9-.081Z"
-        fill="#24285B"
-      />
-    </Svg>
-  );
-}
