--- src/Views/Home/Illustrations/SportIllustration.tsx
@@ -1,153 +1,0 @@
-import Svg, { Path } from "react-native-svg";
-
-export interface SportIllustrationProps {
-  width?: string;
-  height?: string;
-}
-
-export default function SportIllustration({
-  width = "100%",
-  height = "100%"
-}: SportIllustrationProps): JSX.Element {
-  return (
-    <Svg width={width} height={height} viewBox="0 0 216 156">
-      <Path
-        opacity="0.3"
-        d="M80.7592 17.4071C54.719 23.9836 34.1415 44.7639 28.1001 71.4937C23.6097 91.3827 26.9428 115.709 57.757 119.336C123.899 127.123 194.352 122.458 202.801 90.2211C211.249 57.9843 194.833 10.339 157.491 6.91559C136.225 4.97339 104.305 11.433 80.7592 17.4071Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M86.041 55.7243L73.476 57.569L73.7985 60.047L86.3635 58.2023L86.041 55.7243Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M74.4838 55.3398C74.3674 54.4454 73.5916 53.8203 72.7511 53.9437C71.9106 54.0671 71.3236 54.8922 71.44 55.7867L72.2513 62.0211C72.3677 62.9155 73.1435 63.5405 73.984 63.4171C74.8245 63.2937 75.4115 62.4686 75.2951 61.5742L74.4838 55.3398Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M88.9671 52.8944C88.8481 52.0005 88.0707 51.3781 87.2306 51.5043C86.3906 51.6305 85.8061 52.4574 85.9251 53.3513L86.7547 59.5821C86.8737 60.476 87.6512 61.0983 88.4912 60.9722C89.3312 60.846 89.9157 60.0191 89.7967 59.1252L88.9671 52.8944Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M97.5116 101.819L86.1986 100.792L86.0481 105.396L97.5174 106.766L97.5116 101.819Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M134.252 59.987C137.551 60.5673 140.93 60.3956 144.162 59.4832C147.395 58.5708 150.407 56.9387 153.001 54.6952C153.001 54.6952 152.879 49.1636 156.797 48.6473C160.714 48.131 161.901 54.5723 158.22 56.8402C154.54 59.1081 145.304 72.2057 131.601 69.821C117.898 67.4362 128.662 56.1027 134.252 59.987Z"
-        fill="#F4A28C"
-      />
-      <Path
-        opacity="0.08"
-        d="M134.252 59.987C137.551 60.5673 140.93 60.3956 144.162 59.4832C147.395 58.5708 150.407 56.9387 153.001 54.6952C153.001 54.6952 152.879 49.1636 156.797 48.6473C160.714 48.131 161.901 54.5723 158.22 56.8402C154.54 59.1081 145.304 72.2057 131.601 69.821C117.898 67.4362 128.662 56.1027 134.252 59.987Z"
-        fill="black"
-      />
-      <Path
-        d="M123.57 19.626L120.138 32.4777L126.336 35.4831L126.069 24.3954L123.57 19.626Z"
-        fill="#F4A28C"
-      />
-      <Path
-        opacity="0.31"
-        d="M126.15 26.3552C124.996 25.9748 123.888 25.4486 122.852 24.7879C122.902 25.9575 123.24 27.0926 123.832 28.0803C124.424 29.068 125.249 29.8739 126.225 30.4178L126.15 26.3552Z"
-        fill="#CE8172"
-      />
-      <Path
-        d="M128.465 18.193C128.465 18.193 129.674 22.5506 129.535 25.6483C129.521 25.9484 129.452 26.2427 129.33 26.514C129.209 26.7854 129.038 27.0286 128.828 27.2297C128.617 27.4307 128.372 27.5857 128.105 27.6856C127.838 27.7855 127.555 27.8284 127.273 27.8118C125.803 27.7134 123.841 27.1296 122.927 24.88L120.942 21.2661C120.615 20.4966 120.52 19.6383 120.672 18.8094C120.823 17.9805 121.213 17.2218 121.787 16.638C123.778 14.3393 127.834 15.5194 128.465 18.193Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M129.073 20.9964L130.566 22.4592C130.656 22.5415 130.724 22.6473 130.763 22.7662C130.803 22.8851 130.813 23.0128 130.792 23.1369C130.771 23.2611 130.72 23.3772 130.643 23.4739C130.567 23.5706 130.469 23.6446 130.357 23.6884L128.818 24.3399L129.073 20.9964Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M127.707 16.1411C127.707 16.1411 130.461 20.1669 124.715 22.9757L121.486 26.0979C121.486 26.0979 118.564 23.0248 118.222 19.46C117.881 15.8952 124.102 10.0318 127.707 16.1411Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M124.715 22.9749C124.715 22.9749 124.356 21.2724 123.17 21.6658C121.984 22.0591 122.487 24.5238 124.177 24.4008L124.715 22.9749Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M141.479 59.4648C137.799 71.4437 121.521 87.2394 121.521 87.2394L108.865 81.2899L98.9758 76.6434C98.987 76.6238 98.9966 76.6032 99.0047 76.5819V76.5143C99.0233 76.4639 99.0446 76.4147 99.0684 76.3668C99.3288 75.7522 99.9653 74.2771 100.77 72.1505L100.799 72.0645C101.088 71.2839 101.377 70.4235 101.736 69.4892C101.736 69.4585 101.736 69.4216 101.765 69.397C103.842 63.4352 106.365 54.6462 106.776 46.146C107.1 39.4344 109.716 35.8388 112.465 33.9151C114.505 32.5204 116.872 31.7583 119.299 31.7147L128.158 36.0171C128.158 36.0171 145.16 47.4982 141.479 59.4648Z"
-        fill="#55789B"
-      />
-      <Path
-        opacity="0.08"
-        d="M108.883 81.2902L98.9934 76.6437C99.0046 76.6241 99.0142 76.6035 99.0223 76.5822C99.0166 76.5601 99.0166 76.5368 99.0223 76.5146C99.0409 76.4642 99.0622 76.415 99.086 76.3671C99.3637 75.6849 100.122 73.8595 100.787 72.1508L100.816 72.0648C101.198 71.0814 101.545 70.1472 101.754 69.4895C101.754 69.4588 101.754 69.4219 101.783 69.3973C103.86 63.4355 106.383 54.6465 106.794 46.1463C107.118 39.4346 109.734 35.8391 112.482 33.9154L120.115 32.4771C120.115 32.4771 133.881 49.9938 119.889 56.3551C108.437 61.5732 109.867 76.9879 108.883 81.2902Z"
-        fill="black"
-      />
-      <Path
-        d="M130.039 110.011L130.114 126.206H125.028C125.028 126.206 123.98 109.685 120.352 98.6223L126.509 96.213C126.509 96.213 130.971 104.24 130.039 110.011Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M110.133 47.0806C119.698 53.485 129.356 39.8281 120.994 31.752C120.78 31.5431 120.555 31.3402 120.329 31.1374C114.803 26.3741 108.553 19.687 95.8626 22.1394C83.1724 24.5917 78.1206 34.0876 78.1206 50.5532C78.1206 50.5532 73.9426 62.1265 80.3369 61.4013C85.5449 60.7866 81.8183 53.1777 81.8183 53.1777C81.8183 53.1777 88.9417 38.771 94.2018 34.1736C97.3672 31.4447 102.818 42.2006 110.133 47.0806Z"
-        fill="#F4A28C"
-      />
-      <Path
-        d="M98.9933 76.6498C98.9933 76.6498 94.6533 85.322 102.049 90.786C109.444 96.25 116.33 92.2365 122.649 111.929L130.259 110.509C130.259 110.509 130.259 92.2365 121.538 87.2519L98.9933 76.6498Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M128.025 123.324C128.025 123.324 132.076 121.56 132.249 124.443L132.296 125.297C132.333 126.013 132.619 126.689 133.097 127.195C133.576 127.701 134.214 128.002 134.888 128.038C135.356 128.059 135.825 128.031 136.288 127.952C136.666 127.889 137.053 127.966 137.384 128.17C137.714 128.375 137.969 128.693 138.105 129.073C138.24 129.453 138.249 129.871 138.129 130.256C138.009 130.642 137.768 130.972 137.446 131.191C136.721 131.621 135.922 131.891 135.096 131.984C129.813 132.783 125.693 133.429 124.883 131.179C124.073 128.93 122.585 124.123 128.025 123.324Z"
-        fill="#55789B"
-      />
-      <Path
-        d="M87.6163 103.57C87.6163 103.57 87.3038 108.235 84.7982 107.098L84.0575 106.766C83.4383 106.477 82.74 106.443 82.0983 106.67C81.4567 106.897 80.9174 107.37 80.5855 107.996C80.3562 108.43 80.1719 108.89 80.0357 109.366C79.9281 109.76 79.6949 110.102 79.3761 110.334C79.0572 110.566 78.6725 110.672 78.2877 110.636C77.903 110.6 77.542 110.423 77.2666 110.135C76.9912 109.847 76.8184 109.467 76.7778 109.059C76.7365 108.179 76.8624 107.299 77.1482 106.471C78.8032 101.081 80.082 96.8771 82.3562 97.0984C84.6304 97.3196 89.3408 98.0326 87.6163 103.57Z"
-        fill="#55789B"
-      />
-      <Path
-        d="M112.696 87.0307C112.696 87.0307 111.539 107.811 102.766 107.811C100.721 107.84 98.677 107.667 96.6615 107.295L96.32 100.257C96.32 100.257 103.339 100.202 100.602 84.228L112.696 87.0307Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M156.675 58.4631L152.793 54.5049L149.315 56.9511L153.423 61.6468L156.675 58.4631Z"
-        fill="#FFD200"
-      />
-      <Path
-        d="M124.137 11.5376C123.856 10.5065 123.35 9.56002 122.662 8.77429C121.973 7.98856 121.121 7.3856 120.173 7.01398C119.221 6.65418 118.21 6.5104 117.203 6.59192C116.196 6.67344 115.217 6.97845 114.328 7.48724C113.819 7.76241 113.361 8.13542 112.98 8.58741C112.646 9.01199 112.387 9.49714 112.216 10.0195C111.773 11.3597 111.744 12.8149 112.132 14.1741C112.52 15.5333 113.305 16.726 114.374 17.5793C118.228 20.4864 125.572 17.106 124.137 11.5376Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M81.8183 53.178C81.4248 53.0551 77.5766 52.3176 77.5766 52.3176L78.1553 48.0583L83.832 49.2875L81.8183 53.178Z"
-        fill="#FFD200"
-      />
-      <Path
-        d="M163.915 132.162H60.4598V133.84H163.915V132.162Z"
-        fill="#FFD200"
-      />
-      <Path
-        d="M154.632 69.8581H158.683C160.097 69.8628 161.461 70.4175 162.517 71.4174C163.572 72.4173 164.246 73.7932 164.412 75.2851L170.818 133.963H161.634L154.632 69.8581Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M59.0941 133.846H170.297C172.292 133.846 174.206 134.688 175.617 136.186C177.027 137.685 177.82 139.717 177.82 141.836C177.82 142.439 177.594 143.018 177.193 143.444C176.791 143.871 176.247 144.11 175.679 144.11H53.6835C53.1157 144.11 52.5711 143.871 52.1695 143.444C51.768 143.018 51.5424 142.439 51.5424 141.836C51.5424 139.717 52.335 137.685 53.7458 136.186C55.1566 134.688 57.07 133.846 59.0651 133.846H59.0941Z"
-        fill="#24285B"
-      />
-      <Path
-        d="M159.731 75.1937H143.516C142.577 75.1937 141.815 76.0027 141.815 77.0007C141.815 77.9987 142.577 78.8077 143.516 78.8077H159.731C160.67 78.8077 161.432 77.9987 161.432 77.0007C161.432 76.0027 160.67 75.1937 159.731 75.1937Z"
-        fill="#FFD200"
-      />
-      <Path
-        d="M162.222 54.6517L151.99 46.6891L150.598 48.7069L160.83 56.6695L162.222 54.6517Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M154.178 45.8255C154.681 45.0972 154.533 44.0749 153.849 43.5423C153.164 43.0096 152.202 43.1683 151.7 43.8966L148.197 48.9734C147.695 49.7018 147.843 50.724 148.527 51.2567C149.211 51.7893 150.174 51.6307 150.676 50.9023L154.178 45.8255Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M165.853 54.9056C166.355 54.1773 166.208 53.1551 165.523 52.6224C164.839 52.0898 163.877 52.2484 163.374 52.9768L159.872 58.0536C159.37 58.7819 159.517 59.8042 160.202 60.3368C160.886 60.8695 161.848 60.7108 162.351 59.9825L165.853 54.9056Z"
-        fill="#E6E6E6"
-      />
-      <Path
-        d="M156.716 50.6079C157.018 50.5966 157.316 50.687 157.567 50.866C157.627 50.9115 157.678 50.9696 157.716 51.0368C157.754 51.1041 157.779 51.179 157.789 51.2568C157.799 51.3347 157.793 51.4138 157.773 51.4893C157.753 51.5649 157.718 51.6351 157.671 51.6958C157.563 51.7845 157.435 51.8417 157.3 51.8617L155.223 52.3903C155.681 52.7455 156.011 53.2578 156.155 53.8408C156.22 54.132 156.218 54.4358 156.149 54.726C156.079 55.0162 155.944 55.2841 155.755 55.5064C155.558 55.7204 155.301 55.8606 155.022 55.9061C154.744 55.9515 154.459 55.8997 154.21 55.7584C154.027 55.6229 153.871 55.4497 153.752 55.2491C153.633 55.0486 153.552 54.8247 153.516 54.5906C153.37 53.6525 153.491 52.6895 153.863 51.8248C154.39 50.7247 155.663 50.5526 156.716 50.6079Z"
-        fill="#F4A28C"
-      />
-      <Path
-        opacity="0.08"
-        d="M156.716 50.6079C157.018 50.5966 157.316 50.687 157.567 50.866C157.627 50.9115 157.678 50.9696 157.716 51.0368C157.754 51.1041 157.779 51.179 157.789 51.2568C157.799 51.3347 157.793 51.4138 157.773 51.4893C157.753 51.5649 157.718 51.6351 157.671 51.6958C157.563 51.7845 157.435 51.8417 157.3 51.8617L155.223 52.3903C155.681 52.7455 156.011 53.2578 156.155 53.8408C156.22 54.132 156.218 54.4358 156.149 54.726C156.079 55.0162 155.944 55.2841 155.755 55.5064C155.558 55.7204 155.301 55.8606 155.022 55.9061C154.744 55.9515 154.459 55.8997 154.21 55.7584C154.027 55.6229 153.871 55.4497 153.752 55.2491C153.633 55.0486 153.552 54.8247 153.516 54.5906C153.37 53.6525 153.491 52.6895 153.863 51.8248C154.39 50.7247 155.663 50.5526 156.716 50.6079Z"
-        fill="black"
-      />
-    </Svg>
-  );
-}
